Ran benchmarks. The hilbert curve is slower by a factor of ~50x, presumably because of all the computation is requires to simply calculate the index to use. We could try caching the index, but then we might as well store the value in the cache.
I wonder if there are better ways to calculate indices while preserving cache locality? These look promising:
- [A Novel Hilbert Curve for Cache-locality Preserving Loops](http://eprints-dev5.cs.univie.ac.at/5726/1/loops.pdf)
- [Efficient Computation of a Hilbert Curve](https://github.com/becheran/fast-hilbert)

Also, the flat array is slightly slow than the array of arrays. Is that because of compiler optimization? Maybe it's because everything is on the stack? Maybe that liiiiitle bit of computation to flatten a coordinate pair is enough to lose all gains.